added: address fields in clinic entity

+++ PetClinicApp/Source/Modules/Pets/Services/DeletePetService.cs
@@ -1,4 +1,5 @@
-﻿using PetClinicApp.Source.Modules.Pets.Entities;
+﻿using PetClinicApp.Source.Modules.Pets.DTO;
+using PetClinicApp.Source.Modules.Pets.Entities;
 using PetClinicApp.Source.Modules.Pets.Repositories;
 using PetClinicApp.Source.Shared.Errors;
 using PetClinicApp.Source.Shared.Services;
@@ -19,7 +20,7 @@ namespace PetClinicApp.Source.Modules.Pets.Services
         }
 
 
-        public async Task<Pet> ExecuteAsync(long id)
+        public async Task<PetDTO> ExecuteAsync(long loggedUserId, long id)
         {
             var petExists = await petsRepository.Find(id);
 
@@ -28,9 +29,14 @@ namespace PetClinicApp.Source.Modules.Pets.Services
                 throw new AppErrorException("Pet does not exists", HttpStatusCode.NotFound);
             }
 
+            if (petExists.UserId != loggedUserId)
+            {
+                throw new AppErrorException("You do not have permission to perform this action.", HttpStatusCode.Forbidden);
+            }
+
             await petsRepository.Delete(id);
 
-            return petExists;
+            return petExists.ToDTO();
         }
     }
 }
